+++ wp_mobile/src/filters/mod.rs
@@ -1,7 +1,9 @@
+mod media_list_filter;
 mod post_filter;
 mod post_list_filter;
 mod post_type_filter;
 
+pub use media_list_filter::MediaListFilter;
 pub use post_filter::AnyPostFilter;
 pub use post_list_filter::PostListFilter;
 pub(crate) use post_list_filter::compare_posts_by_order;